Added nested loops, inner code as a separate file in csdlib
