+++ ic.txt
@@ -1,15 +1,14 @@
-r(0,0,100,10,1)
-r(0,20,100,10,1)
-r(0,40,100,10,1)
+v(x,0)
+v(w,100)
+v(h,10)
+v(p,0)
 
-r(150,0,100,10,2)
-r(150,20,100,10,2)
-r(150,40,100,10,2)
 
-r(300,0,100,10,3)
-r(300,20,100,10,3)
-r(300,40,100,10,3)
+l(3)
+a(x,150)
+a(p,1)
+v(y,0)
 
-#c(50,25,70,0)
-#c(200,25,70,0)
-#c(350,25,70,0)
+l(4)
+r(x,y,w,h,p)
+a(y,20)
